requirements.txt
openai
imapclient
email
python-dotenv
